fix: warehouse transformer uploader
